refactor: Lecture 폴더 컴포넌트 최적화
