+++ src/components/Lecture/LectureDetail.jsx
@@ -1,53 +1,12 @@
-import { useQuery } from 'react-query';
-import { useSearchParams } from 'react-router-dom';
-import { useSetRecoilState } from 'recoil';
-import { Lecture } from 'api';
 import { LectureInfoBox } from 'components';
 import { fakeLectureInfo } from 'components/placeholderData';
-import { lectureState } from 'app/recoilStore';
-import { isLoginStorage } from 'utils/loginStorage';
+import useLectureQuery from 'hooks/useLectureQuery';
 
 const LectureDetail = () => {
-  const lectures = Lecture();
-  const [searchparams] = useSearchParams();
-  const selectId = searchparams.get('id');
-  const setLectureInfo = useSetRecoilState(lectureState);
-  const isLogin = isLoginStorage();
-  const { data: lecture, isLoading } = useQuery(
-    ['lecture', 'detail', selectId],
-    () => lectures.detail(selectId),
-    {
-      cacheTime: 0,
-      staleTime: 0,
-      enabled: isLogin,
-      onSuccess: (lecture) => {
-        setLectureInfo({
-          selectId: selectId,
-          lectureName: lecture.data.lectureName,
-          professor: lecture.data.professor,
-          semesterList: lecture.data.semesterList,
-          selectedSemester: '선택',
-          satisfaction: 0.5,
-          honey: 0.5,
-          learning: 0.5,
-          team: undefined,
-          homework: undefined,
-          difficulty: undefined,
-          examInfo: '',
-          examType: '선택',
-          examDifficulty: '',
-          content: '',
-        });
-      },
-    }
-  );
-
+  const { Detail } = useLectureQuery();
+  const { data: lecture, isLoading, isLogin } = Detail();
   if (isLoading) return <LectureInfoBox isLogin={true} current={fakeLectureInfo} />;
-  return isLogin ? (
-    <LectureInfoBox isLogin={isLogin} current={lecture.data} />
-  ) : (
-    <LectureInfoBox isLogin={isLogin} current={fakeLectureInfo} />
-  );
+  return <LectureInfoBox isLogin={isLogin} current={isLogin ? lecture.data : fakeLectureInfo} />;
 };
 
 export default LectureDetail;
